Major rewrite and breakage
Work on removing duplicities continues: Get rid of the ugly
document_start.js ifs for each specific functionality separately.
Use levels object for the information on which functionality should be
blocked.

Hence, levels are blocked as they will definitely be different comared
to their old content.

All the settings are still the same. They need to be fixed once levels
are ready.

In the process, several additions and removals for the blocked functions
were done:

* The functions to be blocked are sorted into categories used by Web API
Manager. The hope is that ultimately, JSR will be able to accomplish all
that is possible with Web API Manager (i.e. block whole APIs) but also
allow blocking only a part of the API, e.g. do not block canvas but
block reading from canvas.

* Canvas: originally, only toDataURL was blocked. But the content of the
canvas was still available through other methods. So now it is possible
to also block CanvasRenderingContext2D.prototype.getImageData and
HTMLCanvasElement.prototype.toBlob.

* Block additionaly methods to get performance data.

* GPS data are not filtered anymore. The plan is to bring the blocking
back but right now, I am not convinced the the rounding was done in a
really good manner and I have no way to test the code.

* UA, language and other identifiable characteristics spoofing was
removed. See the paper FP-Scanner: The privacy implications of browser
fingerprint inconsistencies and pages like
https://ghacksuserjs.github.io/TorZillaPrint/TorZillaPrint.html.
Spoofing like that might return in the future once we are sure that we
really help the users.

* The extension detects that CSP prevents wrapping as a result of the
bug https://bugzilla.mozilla.org/show_bug.cgi?id=1267027 in Firefox. The
extension tries to use Firefox-specific exportFunction but so far it
seems that sometimes Firefox does not allow to access page scripts to
the values returned by the exported function.