custom levels: Add the possibility to deselect battery wrapping
Based on https://lukaszolejnik.com/AssessingPrivacyWebStandardsIWPE17.pdf,
IV. USE AND MISUSE OF THE API IN THE WILD - A. Statistics on usage in the wild
that battery status API is mostly used by malicous software and also disabled in most of the browser except Google Chrome, Hence, we protect users by default.

See #47.