Moved shared code, added "shield OFF message"
The majority of the shared code between both versions was moved to common file - http_shield_common, unfortunately, its neccessary to keep messageListener in both versions instead of moving it to the common file, because each browser is sending the response in the different manner and I did not find the way to unify that.

Added message into popup, that says, that Network boundary shield is offline. Message is replacing the "whitelist switch" when the shield is turned off.