Web/docs: be more specific with data collection explanation
Note that in each browser, we technically collect data such as custom
levels and the domains to which they apply. If the user sets up a
browser account, the configuration is stored on the servers of the
user's choice.

The intended meaning is that the extension does not collect any data
about our users that is sent to us.

Nevertheless, keep in mind that GitHub does have statistics, AMO has
statistics (there is probably an opt out in Firefox settings), and most
likely Google and Opera store collect some data for unknown purposes.