Check that XPCNativeWrapper is available during XRAY init
xrayWindow.top !== unwrappedWindow.top ... true in privileged mode in Firefox
typeof XPCNativeWrapper !== "undefined" ... the code bellow depends on XPCNativeWrapper.

The rationale behind testing both is that if the test is true we later do use XPCNativeWrapper, leading to catastrophe if for some reason (e.g. window.top identity not being guaranteed in some other browser != Firefox) that's not actually there. On the other hand, XPCNativeWrapper used to be accessible to unprivileged page scripts as well for some time (not sure if it was still the case in 78, can check later), so the second test alone may not suffice either.