Cascade top document's level to subframes with no explicitly assigned level.
Keep in mind that if a subframe domain has a level assigned that level
is set and not the one of the top level window Origin.

This is a workaround for Github #133

Known bug:

Keep in mind that the code treats some nested iframes incorrectly.
Suppose that there is an iframe from domain C nested in an iframe from
domain B that is iself nested in a visited domain A.

+------------------------------------------------------------+
| visited domain a.example                                   |
|                                                            |
| +--------------------------------------------------------+ |
| | iframe from domain b.example                           | |
| |                                                        | |
| | +----------------------------------------------------+ | |
| | | iframe from domain c.example                       | | |
| | |                                                    | | |
| | +----------------------------------------------------+ | |
| |                                                        | |
| +--------------------------------------------------------+ |
+------------------------------------------------------------+

Suppose that B has a user-defined specific level settings, and C does
not have a user-defined specific level settings. The iframe of domain B
gets the user-defined settings for domain B but the iframe from domain C
is set with the level of domain A.