Bugfix: Populate popup window in inactive tabs
1. Firefox is started with several tabs from previous sessions. The pages
in these tabs were not loaded after the current browser launch.

2. User clicks on one of these tabs.

3. Users quickly click on JShelter icon causing the popup to load.

4.(past) getCurrentSite() returned null due to Missing host permission for
the tab exception.

4.(fixed) The exception is catched and getCurrentSite() is called
  recursievely. Consequently, when the tab is correctly initialized,
  JShelter can populate the content.

Possible downside: If the exception is called in other contexts, we
might cause infinite recursion. As the popup is short-lived and
currently there is no content to iteract with besides global option
button (that causes the pop up to close), we expect that the recursion
does not cause problems.

Suggested solutions: Adding additional parameter to getCurrentSite()
that is set when called recursively. The sleep interval can be longer
with each recursive call. The number of recursive calls can be limited.