Redefine level representation
Levels and tweaks share the same logic and syntax. This allows to reuse
the same UI for tweaks in popup and options_domains for definition of a
new level and editing an existing level.

Additionally, some inconsistences in update mechanism of hardware and
enumerateDevices found.

Update mechanism for new levels provided. But keep in mind that it is
not backwards-compatible so do not use previous versions with the
configuration introduced in this commit. Backup your configuration.