Make slider react to all events not just clicks
See also https://stackoverflow.com/questions/19927813/make-a-checkbox-look-like-a-button-css/19928234:

I'd like to add that you should generally never use onclick on a checkbox or radiobutton since this doesn't cover other methods of toggling (e.g. via keyboard or clicking on a label). Use onchange instead. This is also likely the reason why it doesn't work when you hide the checkbox. –
    Daniel Perván
    Nov 12 2013

Do not blindly refere to a slider at position 0 to be NBS and 1 FPD,
prepare for future sliders that may be above these two sliders.