GPS wrapper reimplemented to use farbling
Previously the each page load generated a position that was used for
each Geolocation reading on that page. So each page load provided a
different position. This was good in terms of yielding different
positions for different domains but the user appeared to hop between
locations for a single domains, which is strange, especially when
combined with the reading of the stationary results of physical sensors.

This changes means that a single domain will keep getting the same
position throughout the session. This will be less confusing to such a
domain and it will also learn less information about the true location.