bugfix: Make sure that farbled APIs produce the same results even when the extension configuration changes
Previously all wrappers shared the same random generator but that means
that the generated random value depended on the order of the call to the
random function.

Hence, if a user enabled or disabled an affected wrapper, other wrappers
might have get a different random number, change its output and possibly
change a fingerprint that should have been stable for the session.