wS-WEBGL: Reimplement farbleGLint
Should the original number be 0, we return the original number but
usually is not 0 (when there is OpenGL supporting hardware).

In other cases JShelter generates a common random number based on
domainHash, and combines the number with other parameters so that
each parameter works a different value. JShelter returns a uniformly
distributed value between 0 and the original value.