Integration tests: NBS_setting test: Popup GIU changed, change the test.
The GUI of the JShelter popup has been changed in the version 0.7. More switchers has been added to the popup. Before, only one switcher for NBS was presented in the popup. Now 3 different switchers are there - one switcher for JS Shield, one for NBS and one for FPD.
Test has to be updated.
Change index of NBS switch in array of all switches in the popup.
Change the CSS selector to be more specific to be able select only NBS switch.