Integration tests: NBS_setting test: Do not test in Google Chrome.
Do not run NBS ON/OFF switch test in Google Chrome.
It is able to show JShelter popup as a website in Mozilla Firefox, but not in the Google Chrome.
To test NBS, JShelter popup has to be opened in a web-browser tab as a webpage, because it is not able to show only popup by Selenium.
When the JShelter popup is not viewable as a webpage in Google Chrome, it is not able to test popup by Selenium automatic tests.