Integration tests: Plugins array test: Ignore known Selenium mistakes.
Ignore the Plugins array test on selected levels.
And update expected values according to documentation.
The Plugin array test has to be ignored on selected levels because Selenium manipulate installed plugins. It seems that error occurs, but added and edited plugins are caused by Selenium.

E.g. On level 0, plugins should be the same as in a web-browser without JShelter. This is fine with manual testing. However, during automatic testing, the installed plugins are changed by Selenium and the test fails because plugins are not the same on level 0 as in a web-browser without JShelter.