Enclose the AFPD wrapper code into try ... catch
The sending code could possibly throw an exception (the other side of
the channel is closed). While this would affect the AFPD functionality,
it does not inerferre with the wrapper itself.

It seems to me that it makes more sense to let the page continue in a
good faith that it would not trigger the detector.

Possible improvement: let the user decide what to do.