Unit tests: Fix removing the custom namespace in a JS source file.
Wrappers has custom namespace created by
(function () {...})()
This custom namespace has to be removed in the source file to unit test on this wrapper can be performed.

Custom namespace defined only by "function()" was removed before this commit. In this commit, removing custom namespace defined by "function ()" is added. The difference is a space between "function" and brackets. Now both variants will be removed to unit tests can be performed on the wrapper.