Unit tests: Add tests for ECMA-ARRAYS and ECMA-SHARED.
Add basic unit tests for ECMA-ARRAY and ECMA-SHARED.
Tests check values and function that should be defined.
ECMA-ARRAYS tests are being created because there is a known bug in the wrapper. ECMA-ARRAY tests should catch this bug in the future.
UPDATE: Integrations tests are much more reliable for catching this bug. Use integration tests to test ECMA-ARRAYS, not unit tests.

Known bug: The function window.crypto.getRandomValues(array) expects an array object as a argument. But proxy array object is given by wrapper. This function raises error because of it.