Integration tests: Define the exact order in which the tests are performed.
In exceptional cases, the tests may affect each other. One test may modify the environment and the other tests will be affected.

Therefore, the exact order in which the tests were performed was established. The method of experiments determines the order of tests so that they do not explain unexplained failures.