Integration tests: Update the accuracy of time.
Time value spooling has changed on level 2. Previously, the value was rounded to the specified number of places. Rounding is no longer used, but noise is added to the real value.

In level 2 tests, accuracy can no longer be checked for the specified number of places. Therefore, the precision is set back to exact values, because the noise values look like exact (non-rounded) values.