Integration tests: Rename JSR to JShelter on the Chrome Options page.
The name of the extension has been changed from JavaScript Restrictor to JShelter in the manifest. This is also reflected on the page with all the extensions in Google Chrome.

The tests were searched for the JavaScript Restrictor on the extensions page in Google Chrome. They will now search for JShelter.