Canvas farbling: Replace mash hash with CRC
This code is about 10-times faster.

Note that the CRC algorithm was not checked that it produces correct
standardized values.

Note that the code is optimised so that there are no checks on input
data; data should be an iterable one-byte-long of integers.

The original Mash that was used to initialize the seed was not
cryptographicaly safe. Neither is CRC16. The attacker can render its
representation of the image for fingerprinting but CRC16 should detect
the differences that the attacker wants to use for fingerprinting. So
the expectation is that attackers cannot create a colliding canvas
content because they do not know the CRC, if the knew, they do not need
to read the canvas in the first place.