Let Alea produce sequences that are not reversible to the attacker
As the PRNG algorithm is publicly known, the attacker could try to
reveal the current state of the PRNG and consequently guess future
pseudo random numbers.

This commit mixes in domain hash so that the attacker would need to know
the domainHash to reconstruct the PRNG state. As we try to keep
domainHash secret from the observers, this commit should make the
guessing of the future PRNG impossible.