wS-WEBA: Keep information on farbled AudioBuffer channels for a limited time
The aim of this change is to prevent the accumulation of memory. Hence,
the stored time might need a (considerable) tweaking.

The disadvantage is that if the page calls the same API for the farbled
data after it was removed, it will get farbled once more. 5 minutes
seems to be a very long time for fingerprinter to repeat the access.