Farbling: Use eTLD+1 instead of origin to generate hash
This means that the same data are farbled on all sites of the same
entity the same way. Of course, there must be a common suffix. So

example.com
example.org

are farbled differently even if they both share the same IP address.

We follow the lead of Brave that behaves the same way. The raitonales
are:

* Should the page detect bots have similar security measures, it will
  detect the same fingerprint and flagging JShelter users should be more
  rare.
* The site owner has other means to track users like cookies so that it
  makes sense to follow Brave.

Possible future improvement: Let the user decide if the hash should
depend on full domain name, origin, or site.