options: Explaint the purpose of JSS custimization
From the UX report: When the toolbar menu is open while on the settings page, the JavaScript
Shield appears to be both off and on. As a result the user may worry that this extension
doesn’t work or that it’s so confusing that they can’t determine whether it’s on or not.

I do not see how to clarify further.