Fix 492873b8: Do not overwrite default level during install
There is a race condition in populating levels and running update.js. In
6279d121, checkAndSaveConfig() is called during every initialization.
492873b8 started to look at the dynamically populated levels. If
the __default__ level is set to a custom level, the level might not be
found in levels if they are not populated, yet.

Unfortunatelly, several users run into the issue, see Pagure #102 and #103
for more details. There is no way to restore the previous state.

Related issues https://pagure.io/JShelter/webextension/issue/102 and
https://pagure.io/JShelter/webextension/issue/103.