Replace Worker polyfill by code that makes Workers unoperable
The previous Workers polyfills did not work well, see https://pagure.io/JShelter/webextension/issue/80.

The previous code was supposed to reate a new scope with this binded to
window. Inside the call, the polyfill launches the code that is
controlled by a possible attacker (downloaded by XHR).

In Firefox, the eval call was executed in strict mode but also in the
privileged context. Chrome wrapper was broken.