Fix popup sometimes not correctly loading current level
In some cases (pretty consistently on chrome), the popup wasn't getting
initialized with the correct level. This was caused by not supplying
the correct type to tabUpdate, it expects a tab object, not a url
string. When the url wasn't saved in tab_urls, empty level was returned.
Changed the onConnect handler to get the correct level directly.