Add Optimized farbling level tweak
This commit adds a new adjustable level for WebAssembly farbling.
The setting is added as a new wrapping group without wrappers. To make
it work correctly, tweaks_gui had to be adjusted to not display hits for
that tweak (there is nothing to track) and to not display the default
off level 'Unprotected' since we want to show 'Disabled' as there is no
protection in this case.

On Chrome, this setting has 3 levels - off, on without CSP modifications
and on with CSP modifications. On Firefox, CSP modifications aren't
needed, so 2 levels - off and on - are used. This tweak has been enabled
(without CSP on Chrome) in the Recommended level.