integration t: Improve checks in webaudio tests
Previous version tested only if the output is or is not equal. Hence, a
small modification in a single value was sufficient for the success of
the tests. This commit evaluates the changes in depth.

Note that some changes are probabilistics, e.g. the value changes with
0.5 probability. This commit introduces an arbitrary chosen threshold of
0.7 same values. Such threshold might fail if by chance not enough
values changed. The commited settings did not provide any false error
during my experiments but the value might possibly need tweaking in the
future.