Generating wrapping code in content scripts
Previously, the code was generated and cached in background. However,
the bachelor thesis of Martin Zmitko
(https://www.vut.cz/en/students/final-thesis/detail/147218) revealed
that there is a huge performance hit. So this commit optimizes page load speeds.

This commit moves generation of the wrappers from background scripts
into content scripts. Most of the time (see the bachelor thesis of
Martin Zmitko), the configuration has to be sent from background
scripts using syncMessage, which, for requests of the original size (~700kB for
Recommended level with FPD), took more than 150ms. In comparison, the
process of code generation including wrapper definition loading does not
exceed 30ms. All measurements in the thesis were made on Chrome. Because wrapping
occurs during early injection on Firefox, the time savings are not as
large, but still apparent as the syncMessage request has to always finish.
Before it finishes, content script and page load is paused.

* Move necessary files between content and page scripts
* Move FPD level initialization in fp_code_builders.js to new background script fp_levels.js
* Injection code wrapping is now called dynamicaly in document_start.js, remove it from levels.js and level_cache.js
* getContentConfiguration now waits for FPD level initialization as well as JSS level initialization
* Reduce size of page configuration for faster transfer and injection times by sending just what's neccessary and using arrays instead of dicts

Additionally, the incoming move to Manifest v3 will make the original
idea to save generated code in background moot as the services workers
would be deactivated often.

This commit is based on the original commit by Martin Zmitko in his
thesis.