level tweaks HTML: Rearrange the description elements to give more clarity
Previously, the user saw:

The name of current settings
Description of the settings (if available)
General description of the group

This was confusing as we first explained specifics and only later gave
the user generic description of the group needed to understand the
specifics.

This commit changes the flow of information to:

General description of the group
The name of current settings
Description of the settings (if available)

Moreover, this flow is the same as appear in levels.js. So the developer
sees the information in the same order as users do.