website: Replace references to white lies by little lies
The extension does uses little lies. The only exception is the question
that mentioned farbling and white lies. I added little lies and kept
white lies so that people finding or remembering the term white lies get
the explanation what it is. The decision to keep reference to white lies
in one question can be reconsidered in the future.