Add benchmark tests used in the optimization blog post
The same tests were also used by Martin Zmitko in his bachelor thesis.
However, the 3D Canvas test was missing in the submitted files so that
was added by me and might be different to the test in the thesis.

I also added the logging of median that I used in the blog post.