window.name: Do not clear during first page load
This closes https://pagure.io/JShelter/webextension/issue/116
This finally solves issues like
https://github.com/polcak/jsrestrictor/issues/156,
https://github.com/polcak/jsrestrictor/issues/168,
https://pagure.io/JShelter/webextension/issue/5,
https://github.com/polcak/jsrestrictor/issues/138

previousUrl in the onUrlChanged listener was undefined while JShelter
cleared the window.name property. The browser was creating a new window
at that time. So the reCaptcha script managed to set window.name before
that window navigated to the initial page. The script might have
propagated that same string in the URL or in a different way. Also, I
guess that Firefox does not have anything special for reCaptcha so
Firefox likely behaves the same. Hence, clearing window.name when
previousUrl is undefined does not make sense.