Fix oversending of FPD messages of HTMLElement.prototype.offset*
(offsetHeight and offsetWidth)

Other FPD wrappers limit the number of messages to 1000 messages per
wrapped property. However, these two special wrappers did not have that
limit. On some pages, like
https://www.gigabyte.com/Motherboard/B550M-AORUS-ELITE-rev-10-11-12#kf
from https://pagure.io/JShelter/webextension/issue/111, the number of
messages was so high that it made the whole browser not responsive.

We need to use a shared counter as distinct wrappers are created for
different elements.