generate_fpd.sh:  Use GNU sed if available
See https://pagure.io/JShelter/webextension/issue/131 and https://pagure.io/JShelter/webextension/pull-request/132

The script prefers gsed if available, if not, it checks uname and if it
detects Mac OS (Darwin) or FreeBSD, it provides a warning that the user
should install GNU sed. It tries sed anyway in the hope that it will
work.