Revert debugging code removal from d8e1bc6ca0080738cc39e37fbb2a9b0314fe3b22
I am not sure why d8e1bc6ca0080738cc39e37fbb2a9b0314fe3b22 commented out
a line of code used to debug the duration of wrapper calls. As
jshelter_debug_timestamp is later used the extension was not usable in debug mode.
It seems that the code works as intended so I am reverting the comment.
However, I did not test thoroughly as the code is not propagated to the
release for public.