Avoid web accessible resources in Chrome
Originally, code_builders.js used WebAssembly.instantiateStreaming() to
inject the assembly code that is supposed to improve performance. See
https://jshelter.org/optimizations/ for more information about the improvements.
WebAssembly.instantiateStreaming() is faster according to the docs at
https://developer.mozilla.org/en-US/docs/WebAssembly/Reference/JavaScript_interface/instantiate_static

"""
Warning: This method is not the most efficient way of fetching and
instantiating Wasm modules. If at all possible, you should use the newer
WebAssembly.instantiateStreaming() method instead, which fetches,
compiles, and instantiates a module all in one step, directly from the
raw bytecode, so doesn't require conversion to an ArrayBuffer.
"""

However WebAssembly.instantiateStreaming() needs a web accessible
resource (WAR) that contains the Web Assembly code.
https://github.com/TecharoHQ/anubis/issues/38#issuecomment-2744603744
contains code that is able to reliably detect that a user has JShelter
installed, even if its protection is off for that site.

Background of the issue:

Web accessible resources are files that are distributed with the
extension and their intended purpose is that they can be accessed by
scripts injected to the page by the extension. In Firefox, the URL of
such a resource looks like moz-extension://[hexa string unique to that
Firefox profile and extension]/warpath. As the extension gets different
hexa string in each install it is impossible to guess the current one.
However, Chrome-based browsers (except Brave) suffer from a bug that the
path is predictable. Such a bug was described in several papers like
https://dl.acm.org/doi/pdf/10.1145/3029806.3029820. We thought that the
bug is long fixed. However, it is not. You can find more on the issue at
https://browserleaks.com/chrome.

https://www.fit.vut.cz/person/polcak/public/jsr/wasm_war.html proves
that JShelter is detectable with this method.

However, such a possibility breaks promises made in our threat model
https://jshelter.org/threatmodel/. So this patch employs a different
path to include web assembly code for Chromium-based users. The web
assembly code in embeded to a JS file during the build time so that no
WAR is needed. Such approach has two drawbacks: the size increases a
little bit as the web assebly code is base64 encoded and
WebAssembly.instantiate is slower.

There is no change for Firefox users.