Remove performance tests
They depend on old lighthouse which in turn dependds on old tar-fs that
contains a security vulnerability.

The current code is not compatible with the new lighthouse.

As I have not run the performance in a while and do not plan to do so
soon, I am removing as outdated. These tests could be bring back but
need to be updated.