Handle pagure issue166
https://pagure.io/JShelter/webextension/issue/166

Show a slightly different version of a message instructing users that
use Chromium-based browsers to activate User Scripts or Developer Mode.

Once all (or most) users are on Chromium 138 or greater, return to the
issue and remove this commit and the old strings and change options.html
and popup.html accordingly.